count: fix a word to fit in context
A word of quick quiz in `Parallel Counting Performance` subsection does
not fit with its context.  It says `` `count_stat.c` row of Table
`Statistical Counter Performance on Power-6` shows that the update side
scales linearly with the number of threads.''.  However, the table shows
read performance under 1 core and 32 cores, not update performance under
different number of cores.  Also, the context says about read-side
performance, not update-side performance.  This commit fix the word to
fit in the context by changing the word to read-side performance.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>