whymb: remove ARM's short multi processor history description
A description about ARM processor says that history of multi processor
ARM CPU is about five years.  However, the sentence was written in 2010.
Now, multi processor ARM CPU is common and its history is more than a
decade.  The sentence can be modified to say ten years rather than five
years.  However, it would be better to simply remove the sentence rather
than updating the word every year.
Moreover, purpose of the sentence is to say about ARM's short multi
processor history.  However, ten years in industry may sufficient to be
a long time to be mature though it can be arguable.  So, removal of the
sentence may would not harm the meaning of the sub section.
For the reason, this commit removes the sentence.

[1] http://www.arm.com/about/newsroom/5346.php

Signed-off-by: SeongJae Park <sj38.park@gmail.com>