advsync: fix wrong reference to section ``MMIO write barrier''
A sentence references ``MMIO write barrier'' subsection though the
section does not exists.  Unfortunately, there is no section for MMIO
write barrier though appendix ``Why Memory Barriers?'' says about
mmiowb() subtly.  For the purpose of the sentence, it could be
alternated with reference to ``Device operations'' subsection because it
describes about the exceptional case.  For the reason, this commit
updates it to the reference to ``Device Operations'' subsection.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>