advsync: use latex reference feature consistently
References to other sections in `Advanced Synchronization` chapter use
latex reference feature and section name direct citation inconsistently.
Because most other parts are using latex reference feature and it is
easier for document maintainance, it would be better to use latex
reference feature only.  For the reason, this commit enforces latex
reference feature to the references.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>