formal: Rearrange promela sample code location
`Formal Verification` was a section in analysis chapter in initial
version.  After that, it changed to an appendix[1], then promoted to a
chapter[2] now while the analysis chapter has merged into debugging
chapter[3].  However, code samples for formal verification exist under
`CodeSamples/appendix/formal/` and `CodeSamples/analysis/` directory.
It is unnecessary duplicate and ambiguous directory hierarchy.  This
commit removes the unnecessarily dulicated files and changes the name of
the directory.

[1] commit 8a7e0cb902e9, ("Move the formal-verification sections to a
    new appendix.")
[2] commit 87a645d2aa0b ("Promote formal-methods appendix to a
    chapter.")
[3] commit e0c08f843331 ("Merge the analysis chapter into debugging.")

Signed-off-by: SeongJae Park <sj38.park@gmail.com>