intro: Polish a sentence for MIPS graphs
This commit polishes a sentence for MIPS graphs, which is updated by
commit "cpu: Update MIPS graphs, adding past five years of CPUs"
(71d8328f3d18).  The sentence is saying that the updated graph is
showing past three decades, though it is now showing past forty years
and the sentence also says so at last.  This commit polishes the
sentence to say about the fourty years.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>