howto: Provide a hyperlink to 'submitting-patches.html'
In 'howto' chapter, we point users to 'SubmittingPatches' file in the
Linux source tree to let them know the detailed contribution process.
The file is now converted to a '.rst' file and the rendered 'html' page
is available at 'kernel.org'.  Using the link to the page would be
easier for the users, compared to retrieving the Linux source tree and
then reading the 'SubmittingPatches' file, which now only directs reader
to the '.rst' file.  This commit therefore makes this change.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>